added params to test texture repeating/mirroring for gouraud example (repeating/mirroring doesn't work 100%, but at least there is a way to test it :p) added silviu's contribution to the repository (the samples here worked on a previous version of the engine and they were not ported)

+++ src/bspworld.h
@@ -216,11 +216,11 @@ public: // get-set
 public: // ops
     float ComputeU(CPoint3D& _p)
     {
-        return (_p.GetX() * mUAxis.GetX() + _p.GetY() * mUAxis.GetY() + _p.GetZ() * mUAxis.GetZ() + mUOffset) / (float)mTex->GetWidth();
+        return (_p.GetX() * mUAxis.GetX() + _p.GetY() * mUAxis.GetY() + _p.GetZ() * mUAxis.GetZ() + mUOffset) / mTex->GetWidth();
     }
     float ComputeV(CPoint3D& _p)
     {
-        return (_p.GetX() * mVAxis.GetX() + _p.GetY() * mVAxis.GetY() + _p.GetZ() * mVAxis.GetZ() + mVOffset) / (float)mTex->GetHeight();
+        return (_p.GetX() * mVAxis.GetX() + _p.GetY() * mVAxis.GetY() + _p.GetZ() * mVAxis.GetZ() + mVOffset) / mTex->GetHeight();
     }
 
 protected:
